Update bot.py
Better user responsive way to do this instead of updating the message after a user reacts and 5 minutes passes. There could be a better way to do it instead of sleeping for 300 seconds but this works fine.

+++ bot.py
@@ -15,13 +15,14 @@ async def votekick(ctx, lol: discord.Member):
         return
     msg = await ctx.reply(f"kick {lol.mention}? (3 votes required)")
     await msg.add_reaction("👍")
+    time.sleep(300)
+    await msg.edit(content=f"the votekick for {lol.mention} has expired after 5 minutes.")
+
 @bot.event 
 async def on_reaction_add(reaction, user):
     expiresAt = int(time.mktime(reaction.message.created_at.timetuple())+300)
     timeNow = int(time.time())
-    if timeNow >= expiresAt:
-        return await reaction.message.edit(content=f"this votekick has expired.")
-    if reaction.count == 4 and reaction.emoji == "👍":
+    if reaction.count == 4 and reaction.emoji == "👍" and timeNow >= expiresAt:
         await reaction.message.delete()
         try:
             await lol.kick()
